Add context lost flags section
For some LPI states it may not be possible to use the timer for
wakeup events as the timer's context is lost on entry to the LPI
state. Specify that the "Arch. Context Lost Flags" LPI field must
be used to indicate whether the timer loses context, or not, to
ensure that the OS can take that information into account when
choosing an LPI state with regard to available wakeup events.

Signed-off-by: Andrew Jones <ajones@ventanamicro.com>